+++ test/integration/backward_compatible/serial/CompactRestartTest.js
@@ -18,10 +18,7 @@
 const TestUtil = require('../../../TestUtil');
 const CompactUtil = require('../parallel/serialization/compact/CompactUtil');
 const RC = require('../../RC');
-const { Predicates, IllegalStateError } = require('../../../../lib');
-const { UuidUtil } = require('../../../../lib/util/UuidUtil');
-const sinon = require('sinon');
-const sandbox = sinon.createSandbox();
+const { Predicates } = require('../../../../lib');
 
 let COMPACT_ENABLED_ZERO_CONFIG_XML = `
     <hazelcast xmlns="http://www.hazelcast.com/schema/config"
@@ -100,29 +97,5 @@ describe('CompactRestartTest', function() {
         // Perform a query to make sure that the schema is available on the cluster
         const values = await map.valuesWithPredicate(Predicates.sql('INT32 == 42'));
         values.size().should.be.eq(1);
-    });
-
-    it('should throw IllegalStateError exception, because of schema could not be replicated in the cluster', async function() {
-        const fakeResult = new Set();
-        sandbox.replace(UuidUtil, 'convertUUIDSetToStringSet', sandbox.fake.returns(fakeResult));
-
-        const client = await testFactory.newHazelcastClientForSerialTests({
-            clusterName: cluster.id,
-            serialization: {
-                compact: {
-                    serializers: [new CompactUtil.FlexibleSerializer([FieldKind.INT32])]
-                }
-            },
-            properties: {
-                'hazelcast.client.schema.max.put.retry.count': 1,
-                'hazelcast.client.invocation.retry.pause.millis': 100
-            }
-        });
-        const map = await client.getMap(mapName);
-        const error = await TestUtil.getRejectionReasonOrThrow(async () => {
-            await map.put(1, new CompactUtil.Flexible({INT32: {value: 42}}));
-        });
-        error.should.be.instanceOf(IllegalStateError);
-        error.message.includes('cannot be replicated in the cluster, after').should.be.true;
     });
 });
